SECS implementation
Added ability to interpolate the magnetic fields using the Spherical Elementary Current Systems method. However, I did some comparisons between the SECS interpolation and the standard natural neighbour interpolation and they (a) weren't all the different and (b) the natural neighbor actually outperformed the SECS in some instances when comparing to a known mag site. So I've included the secs function in the repo, but I still use the regular interp in GIC_MAIN.