Update interpolate_Z.m
Cleaned up some of the variable names. Also changed the Z variable to be from 1:end and 2:end instead of 2:end and 2:end-1. This is the correct way to do it based on the way the Fourier coefficients are binned (but it doesn't effect the answer except at the DC frequency).

Also added option to output the interpolated Z at the query points without negative frequencies.